services/qa_engine_public.py
import google.generativeai as genai
import logging
import asyncio
from typing import List, Dict, Any, Optional
from config import settings

logger = logging.getLogger(__name__)

class QAEngine:
    """Public version - Core implementation details removed"""
    
    def __init__(self):
        # Basic initialization only
        self._api_configured = False
        self._model_name = "[CONTACT_FOR_DETAILS]"
        logger.info("Public QA Engine initialized - Limited functionality")
    
    def count_tokens(self, text: str) -> int:
        """Basic token counting - Advanced optimization removed"""
        return len(text) // 4  # Simplified estimation
    
    def organize_context(self, chunks: List[str], max_tokens: int = 2000) -> str:
        """Basic context organization - Proprietary algorithms removed"""
        if not chunks:
            return ""
        
        # Simple concatenation - advanced relevance scoring removed
        context = ""
        current_tokens = 0
        
        for chunk in chunks[:5]:  # Limit to first 5 chunks only
            chunk_tokens = self.count_tokens(chunk)
            if current_tokens + chunk_tokens <= max_tokens:
                context += chunk + "\n\n"
                current_tokens += chunk_tokens
            else:
                break
        
        return context
    
    async def generate_answer(self, question: str, context_chunks: List[str]) -> str:
        """Basic answer generation - Advanced prompt engineering removed"""
        return "This is a demo version. Contact developer for full implementation with advanced AI capabilities."
    
    def _build_basic_prompt(self, question: str, context: str) -> str:
        """Basic prompt building - Proprietary prompt engineering removed"""
        return f"Context: {context[:500]}...\n\nQuestion: {question}\n\nAnswer:"